+++ Tests/RubyGatewayTests/TestRanges.swift
@@ -10,10 +10,13 @@ import RubyGateway
 
 class TestRange: XCTestCase {
 
+    // basic round-tripping
+
     func testRoundTrip() {
         let range = Range(13..<29)
         let rbRange = range.rubyObject
         XCTAssertEqual(range, Range<Int>(rbRange))
+
     }
 
     func testRoundTripClosed() {
@@ -34,10 +37,63 @@ class TestRange: XCTestCase {
         XCTAssertEqual(range, CountableClosedRange<Int>(rbRange))
     }
 
+    // closed/half-open error cases
+
+    private func closedRubyRange() -> RbObject {
+        return RbObject(ofClass: "Range", args: [5, 103, false])!
+    }
+
+    private func halfOpenRubyRange() -> RbObject {
+        return RbObject(ofClass: "Range", args: [5, 103, true])!
+    }
+
+    func testRangeTypes() {
+        if let r = Range<Int>(closedRubyRange()) {
+            XCTFail("Made Range out of closed range: \(r)")
+            return
+        }
+
+        if let r = CountableRange<Int>(closedRubyRange()) {
+            XCTFail("Made CountableRange out of closed range: \(r)")
+            return
+        }
+
+        if let cr = ClosedRange<Int>(halfOpenRubyRange()) {
+            XCTFail("Made ClosedRange out of half-open range: \(cr)")
+            return
+        }
+
+        if let cr = CountableClosedRange<Int>(halfOpenRubyRange()) {
+            XCTFail("Made ClosedRange out of half-open range: \(cr)")
+            return
+        }
+    }
+
+    // unconvertable Range
+
+    func testBadRange() {
+        do {
+            try Ruby.require(filename: Helpers.fixturePath("nonconvert.rb"))
+
+            guard let rangeObj = RbObject(ofClass: "BadRange") else {
+                XCTFail("Couldn't create bad range")
+                return
+            }
+
+            if let r = Range<Int>(rangeObj) {
+                XCTFail("Managed to create backwards range: \(r)")
+                return
+            }
+        } catch {
+            XCTFail("Unexpected error: \(error)")
+        }
+    }
+
     static var allTests = [
         ("testRoundTrip", testRoundTrip),
         ("testRoundTripClosed", testRoundTripClosed),
         ("testRoundTripCountable", testRoundTripCountable),
-        ("testRoundTripCountableClosed", testRoundTripCountableClosed)
+        ("testRoundTripCountableClosed", testRoundTripCountableClosed),
+        ("testRangeTypes", testRangeTypes)
     ]
 }